Task 1 - Game Of Life/Solution/GameOfLife/GameOfLife.ConsoleApp/InputManager.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace GameOfLife.ConsoleApp
{
    public class InputManager : IInputManager
    {
        public string ReadLine()
        {
            return Console.ReadLine();
        }

        public void Write(string value)
        {
            Console.Write(value);
        }

        public void WriteLine(string message)
        {
            Console.WriteLine(message);
        }
    }
}
